Moved account update functionality from FirebaseContext
